Move preview Buffer class to gl_helpers so that it can be shared
Previously, all three OpenGL preview window types (X11, Wayland,
Wayland "direct") were duplicating this code exactly.
