+++ picamera2/previews/q_gl_picamera2.py
@@ -19,12 +19,7 @@ from OpenGL.GLES2.OES.EGL_image_external import *
 from OpenGL.GLES2.VERSION.GLES2_2_0 import *
 from OpenGL.GLES3.VERSION.GLES3_3_0 import *
 
-from picamera2.previews.gl_helpers import (
-    check_egl_extensions,
-    getEGLNativeDisplay,
-    glEGLImageTargetTexture2DOES,
-    str_to_fourcc,
-)
+from picamera2.previews.gl_helpers import Buffer, check_egl_extensions, getEGLNativeDisplay
 
 from .qt_compatibility import _QT_BINDING, _get_qt_modules
 
@@ -266,94 +261,6 @@ def _get_qglpicamera2(qt_module: _QT_BINDING):
             glUseProgram(self.program_overlay)
             glUniform1i(glGetUniformLocation(self.program_overlay, "overlay"), 0)
 
-        class Buffer:
-            # libcamera format string -> DRM fourcc, note that 24-bit formats are not supported
-            FMT_MAP = {
-                "XRGB8888": "XR24",
-                "XBGR8888": "XB24",
-                "YUYV": "YUYV",
-                # doesn't work "YVYU": "YVYU",
-                "UYVY": "UYVY",
-                # doesn't work "VYUY": "VYUY",
-                "YUV420": "YU12",
-                "YVU420": "YV12",
-                "NV12": "NV12",
-            }
-
-            def __init__(self, display, completed_request, max_texture_size):
-                picam2 = completed_request.picam2
-                stream = picam2.stream_map[picam2.display_stream_name]
-                fb = completed_request.request.buffers[stream]
-
-                cfg = stream.configuration
-                pixel_format = str(cfg.pixel_format)
-                if pixel_format not in self.FMT_MAP:
-                    raise RuntimeError(f"Format {pixel_format} not supported by QGlPicamera2 preview")
-                fmt = str_to_fourcc(self.FMT_MAP[pixel_format])
-                w, h = (cfg.size.width, cfg.size.height)
-                if w > max_texture_size or h > max_texture_size:
-                    raise RuntimeError(f"Maximum supported preview image size is {max_texture_size}")
-                if pixel_format in ("YUV420", "YVU420"):
-                    h2 = h // 2
-                    stride2 = cfg.stride // 2
-                    # fmt: off
-                    attribs = [
-                        EGL_WIDTH, w,
-                        EGL_HEIGHT, h,
-                        EGL_LINUX_DRM_FOURCC_EXT, fmt,
-                        EGL_DMA_BUF_PLANE0_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE0_OFFSET_EXT, 0,
-                        EGL_DMA_BUF_PLANE0_PITCH_EXT, cfg.stride,
-                        EGL_DMA_BUF_PLANE1_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE1_OFFSET_EXT, h * cfg.stride,
-                        EGL_DMA_BUF_PLANE1_PITCH_EXT, stride2,
-                        EGL_DMA_BUF_PLANE2_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE2_OFFSET_EXT, h * cfg.stride + h2 * stride2,
-                        EGL_DMA_BUF_PLANE2_PITCH_EXT, stride2,
-                        EGL_NONE,
-                    ]
-                    # fmt: on
-                elif pixel_format == "NV12":
-                    h2 = h // 2
-                    # fmt: off
-                    attribs = [
-                        EGL_WIDTH, w,
-                        EGL_HEIGHT, h,
-                        EGL_LINUX_DRM_FOURCC_EXT, fmt,
-                        EGL_DMA_BUF_PLANE0_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE0_OFFSET_EXT, 0,
-                        EGL_DMA_BUF_PLANE0_PITCH_EXT, cfg.stride,
-                        EGL_DMA_BUF_PLANE1_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE1_OFFSET_EXT, h * cfg.stride,
-                        EGL_DMA_BUF_PLANE1_PITCH_EXT, cfg.stride,
-                        EGL_NONE,
-                    ]
-                    # fmt: on
-                else:
-                    # fmt: off
-                    attribs = [
-                        EGL_WIDTH, w,
-                        EGL_HEIGHT, h,
-                        EGL_LINUX_DRM_FOURCC_EXT, fmt,
-                        EGL_DMA_BUF_PLANE0_FD_EXT, fb.planes[0].fd,
-                        EGL_DMA_BUF_PLANE0_OFFSET_EXT, 0,
-                        EGL_DMA_BUF_PLANE0_PITCH_EXT, cfg.stride,
-                        EGL_NONE,
-                    ]
-                    # fmt: on
-
-                image = eglCreateImageKHR(display, EGL_NO_CONTEXT, EGL_LINUX_DMA_BUF_EXT, None, attribs)
-
-                self.texture = glGenTextures(1)
-                glBindTexture(GL_TEXTURE_EXTERNAL_OES, self.texture)
-                glTexParameteri(GL_TEXTURE_EXTERNAL_OES, GL_TEXTURE_MAG_FILTER, GL_LINEAR)
-                glTexParameteri(GL_TEXTURE_EXTERNAL_OES, GL_TEXTURE_MIN_FILTER, GL_LINEAR)
-                glTexParameteri(GL_TEXTURE_EXTERNAL_OES, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE)
-                glTexParameteri(GL_TEXTURE_EXTERNAL_OES, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE)
-                glEGLImageTargetTexture2DOES(GL_TEXTURE_EXTERNAL_OES, image)
-
-                eglDestroyImageKHR(display, image)
-
         def set_overlay(self, overlay):
             if not self.picamera2.camera_config:
                 raise RuntimeError("Camera must be configured before setting overlay")
@@ -395,7 +302,7 @@ def _get_qglpicamera2(qt_module: _QT_BINDING):
 
                 if self.picamera2.verbose_console:
                     print("Make buffer for request", completed_request.request)
-                self.buffers[completed_request.request] = self.Buffer(
+                self.buffers[completed_request.request] = Buffer(
                     self.egl.display, completed_request, self.egl.max_texture_size
                 )
